# apps/web/notes/openapi.json
{
  "components": {
    "schemas": {
      "ChoiceRequest": {
        "additionalProperties": false,
        "description": "Request payload for resolving a player choice.",
        "properties": {
          "choice_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Choice Id"
          },
          "payload": {
            "title": "Payload"
          }
        },
        "title": "ChoiceRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "contact": {
      "email": "tangldev@storytan.gl",
      "name": "TanglDev",
      "url": "https://github.com/derekmerck/storytangl"
    },
    "description": "RESTful endpoints for StoryTangl",
    "title": "StoryTangl",
    "version": "3.7.2"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/story/do": {
      "post": {
        "description": "Resolve a player choice and return the updated runtime envelope.",
        "operationId": "do_story_action_story_do_post",
        "parameters": [
          {
            "description": "Response rendering profile.",
            "in": "query",
            "name": "render_profile",
            "required": false,
            "schema": {
              "default": "raw",
              "description": "Response rendering profile.",
              "title": "Render Profile",
              "type": "string"
            }
          },
          {
            "example": "AXxngZiGzAqUJqMqzbMmtQ==",
            "in": "header",
            "name": "X-API-Key",
            "required": true,
            "schema": {
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChoiceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Do Story Action",
        "tags": [
          "Story"
        ]
      }
    },
    "/story/drop": {
      "delete": {
        "description": "End the user's active story and optionally archive the ledger.",
        "operationId": "reset_story_story_drop_delete",
        "parameters": [
          {
            "description": "Retain the ledger if true.",
            "in": "query",
            "name": "archive",
            "required": false,
            "schema": {
              "default": false,
              "description": "Retain the ledger if true.",
              "title": "Archive",
              "type": "boolean"
            }
          },
          {
            "description": "Response rendering profile.",
            "in": "query",
            "name": "render_profile",
            "required": false,
            "schema": {
              "default": "raw",
              "description": "Response rendering profile.",
              "title": "Render Profile",
              "type": "string"
            }
          },
          {
            "example": "AXxngZiGzAqUJqMqzbMmtQ==",
            "in": "header",
            "name": "X-API-Key",
            "required": true,
            "schema": {
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reset Story",
        "tags": [
          "Story"
        ]
      }
    },
...
    "/story/info": {
      "get": {
        "description": "Return runtime status details for the active story.",
        "operationId": "get_story_info_story_info_get",
        "parameters": [
          {
            "description": "Response rendering profile.",
            "in": "query",
            "name": "render_profile",
            "required": false,
            "schema": {
              "default": "raw",
              "description": "Response rendering profile.",
              "title": "Render Profile",
              "type": "string"
            }
          },
          {
            "example": "AXxngZiGzAqUJqMqzbMmtQ==",
            "in": "header",
            "name": "X-API-Key",
            "required": true,
            "schema": {
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Story Info",
        "tags": [
          "Story"
        ]
      }
    },
...
    "/story/story/create": {
      "post": {
        "description": "Create a story session and return the initial runtime envelope.",
        "operationId": "create_story_story_story_create_post",
        "parameters": [
          {
            "description": "World template to instantiate",
            "in": "query",
            "name": "world_id",
            "required": true,
            "schema": {
              "description": "World template to instantiate",
              "title": "World Id",
              "type": "string"
            }
          },
          {
            "description": "Optional story label",
            "in": "query",
            "name": "story_label",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional story label",
              "title": "Story Label"
            }
          },
          {
            "description": "Initialization mode: LAZY or EAGER",
            "in": "query",
            "name": "init_mode",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Initialization mode: LAZY or EAGER",
              "title": "Init Mode"
            }
          },
          {
            "description": "Response rendering profile.",
            "in": "query",
            "name": "render_profile",
            "required": false,
            "schema": {
              "default": "raw",
              "description": "Response rendering profile.",
              "title": "Render Profile",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": true,
            "schema": {
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Story",
        "tags": [
          "Story"
        ]
      }
    },
    "/story/update": {
      "get": {
        "description": "Return the runtime envelope with ordered fragments.",
        "operationId": "get_story_update_story_update_get",
        "parameters": [
          {
            "description": "Response rendering profile.",
            "in": "query",
            "name": "render_profile",
            "required": false,
            "schema": {
              "default": "raw",
              "description": "Response rendering profile.",
              "title": "Render Profile",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Inclusive starting step; defaults to 0 (full history).",
            "in": "query",
            "name": "since_step",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Inclusive starting step; defaults to 0 (full history).",
              "title": "Since Step"
            }
          },
          {
            "example": "AXxngZiGzAqUJqMqzbMmtQ==",
            "in": "header",
            "name": "X-API-Key",
            "required": true,
            "schema": {
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Story Update Story Update Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Story Update",
        "tags": [
          "Story"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Reference API",
      "url": "https://localhost:8000/api/v2"
    }
  ]
}